services/siya/shipments_api/serializers/shipment_type.py
from rest_framework import serializers
from core_models.models import ShipmentType


class ShipmentTypeSerializer(serializers.ModelSerializer):
    class Meta:
        model = ShipmentType
        fields = "__all__"
