services/siya/requirements.txt
asgiref==3.11.0
attrs==25.4.0
blinker==1.9.0
click==8.3.1
Django==5.2.10
djangorestframework==3.16.1
drf-spectacular==0.29.0
Flask==3.1.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
mysqlclient==2.2.7
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
Werkzeug==3.1.5
gunicorn==22.0.0

django-extensions
